rename testutil/audit to testutil/testaudit
